requirements.txt
acme==1.21.0
apturl==0.5.2
asgiref==3.6.0
attrs==21.2.0
awscli==1.22.34
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
botocore==1.23.34
Brlapi==0.8.3
brz-etckeeper==0.0.0
CacheControl==0.12.11
cachetools==5.3.0
certbot==1.21.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.0.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
ConfigArgParse==1.5.3
configobj==5.0.6
cryptography==39.0.1
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.1.7
django-dotenv==1.4.2
djangorestframework==3.14.0
docutils==0.17.1
duplicity==0.8.21
fail2ban==0.11.2
fasteners==0.14.1
filelock==3.12.4
firebase-admin==6.1.0
future==0.18.2
google-api-core==2.11.0
google-api-python-client==2.79.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
html5lib==1.1
httplib2==0.21.0
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
jmespath==0.10.0
josepy==1.10.0
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libvirt-python==8.0.0
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.4
mysqlclient==2.1.1
netifaces==0.11.0
nftables==0.1
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
parsedatetime==2.6
pexpect==4.8.0
Pillow==9.0.1
pipenv==2023.9.8
platformdirs==3.10.0
proto-plus==1.22.2
protobuf==4.22.0
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
pytz==2022.7.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.28.2
requests-toolbelt==0.9.1
roman==3.3
rsa==4.9
s3transfer==0.5.0
scour==0.38.2
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.3
ssh-import-id==5.11
systemd-python==234
terminator==2.1.1
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.14
usb-creator==0.3.7
virtualenv==20.24.5
wadllib==1.3.6
webencodings==0.5.1
xdg==5
xkit==0.0.0
youtube-dl==2021.12.17
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
